LayoutTests/ipc/dispatched-from-to-exposed-to-ipc-testing-api.html

<!doctype html><!-- webkit-test-runner [ IPCTestingAPIEnabled=true ] -->
<title>Test that dispatchedFrom and dispatchedTo are correctly exposed to the IPCTestingAPI</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/ipc.js"></script>
<body>
<p>Test pass if assertions don't cause error</p>
<script>

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

if (window.IPC) {
    assert_equals(IPC.messages.IPCTester_AsyncOptionalExceptionData.dispatchedFrom, "Unknown");
    assert_equals(IPC.messages.IPCTester_AsyncOptionalExceptionData.dispatchedTo, "Unknown");
    assert_equals(IPC.messages.WebPageTesting_IsLayerTreeFrozen.dispatchedFrom, "UI");
    assert_equals(IPC.messages.WebPageTesting_IsLayerTreeFrozen.dispatchedTo, "WebContent");
    assert_equals(IPC.messages.WebPageProxy_CreateNewPage.dispatchedFrom, "WebContent");
    assert_equals(IPC.messages.WebPageProxy_CreateNewPage.dispatchedTo, "UI");
    assert_equals(IPC.messages.NetworkProcessProxy_DidReceiveAuthenticationChallenge.dispatchedFrom, "Networking");
    assert_equals(IPC.messages.NetworkProcessProxy_DidReceiveAuthenticationChallenge.dispatchedTo, "UI");
    assert_equals(IPC.messages.GPUConnectionToWebProcess_CreateRenderingBackend.dispatchedFrom, "WebContent");
    assert_equals(IPC.messages.GPUConnectionToWebProcess_CreateRenderingBackend.dispatchedTo, "GPU");
}

document.write("PASS");
testRunner.notifyDone();
</script>
</body>
